cloud-function/requirements.txt
requests>=2.28.0       # For making HTTP requests
gpxpy>=1.5.0           # For parsing and working with GPX files
SRTM.py>=0.3.7         # For accessing SRTM elevation data (tkrajina/srtm.py)
beautifulsoup4>=4.11.0 # For HTML parsing with BeautifulSoup
